+++ packages/plugin-link-preview/web/src/consts.ts
@@ -24,3 +24,7 @@ export const socialModals = {
   Twitter: 'twitterInsertModal',
   Pinterest: 'pinterestInsertModal',
 };
+
+export const linkPreviewModals = {
+  settings: 'settings',
+};